Category bug

+++ src/main/java/com/ktm/controller/CategoryController.java
@@ -2,9 +2,14 @@ package com.ktm.controller;
 
 
 import com.alibaba.fastjson.JSON;
+import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
 import com.ktm.common.Result;
+import com.ktm.model.Article;
 import com.ktm.model.Category;
+import com.ktm.model.Comment;
+import com.ktm.service.IArticleService;
 import com.ktm.service.ICategoryService;
+import com.ktm.service.ICommentService;
 import com.ktm.vo.CateGoryVo;
 import org.springframework.web.bind.annotation.*;
 
@@ -13,6 +18,7 @@ import javax.servlet.http.HttpServletRequest;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
+import java.util.stream.Collectors;
 
 /**
  * <p>
@@ -24,7 +30,6 @@ import java.util.List;
  */
 @RestController
 @RequestMapping("/manager/category")
-@CrossOrigin
 public class CategoryController {
 
 
@@ -32,8 +37,20 @@ public class CategoryController {
     private ICategoryService categoryService;
 
 
+    @Resource
+    private ArticleController articleController;
+
+
+    @Resource
+    private IArticleService articleService;
+
+    @Resource
+    private ICommentService commentService;
+
+
     /**
      * 查询所有栏目
+     *
      * @return
      */
     @GetMapping("/findAllCategory")
@@ -47,6 +64,7 @@ public class CategoryController {
 
     /**
      * 保存或者更新栏目
+     *
      * @param category
      * @return
      */
@@ -55,29 +73,48 @@ public class CategoryController {
 
         int resultNumber = categoryService.saveOrUpdateCategory(category);
 
-        return Result.ok("success",null,200);
+        return Result.ok("success", null, 200);
     }
 
     /**
      * 删除单个栏目
+     *
      * @param id
      * @return
      */
     @RequestMapping("/deleteCategoryById")
     public Result deleteCategoryById(int id) {
-
-        boolean result = categoryService.removeById(id);
-
-        if (result) {
-            return Result.ok();
+        List<Category> categories = categoryService.list(new QueryWrapper<Category>().eq("parentId", id));
+        if (categories.size() > 0) {
+            return Result.fail("你要删除的栏目包含子栏目，不能删除", null, 500);
         }
 
-        return Result.fail();
+        categoryService.removeById(id);
+
+        List<Article> articles = articleService.list(new QueryWrapper<Article>().eq("categoryId", id));
+        if (articles.size() == 0) {
+            return Result.ok();
+        }
+        articleService.remove(new QueryWrapper<Article>().eq("categoryId", id));
+
+        List<Article> articleList = articleService.list(new QueryWrapper<Article>().eq("categoryId", id));
+        if (articleList.size() == 0) {
+            return Result.ok();
+        }
+        List<Integer> integerList = articleList
+                .stream().map(e -> e.getId()).collect(Collectors.toList());
+
+
+        commentService.remove(new QueryWrapper<Comment>().in("articleid", integerList));
+
+
+        return Result.ok();
     }
 
 
     /**
      * 批量删除栏目
+     *
      * @param ids
      * @return
      */
@@ -86,12 +123,32 @@ public class CategoryController {
 
         List<Integer> list = Arrays.asList(ids);
 
-        boolean result = categoryService.removeByIds(list);
 
-        if (result) {
+        List<Category> categories = categoryService.list(new QueryWrapper<Category>().in("parentId", list));
+
+        if (categories.size() > 0) {
+            return Result.fail("你要删除的某些栏目里包含子栏目，不能删除", null, 500);
+        }
+
+        categoryService.removeByIds(list);
+
+        List<Article> articles = articleService.list(new QueryWrapper<Article>().in("categoryId", list));
+        if (articles.size() == 0) {
+            return Result.ok();
+        }
+        articleService.remove(new QueryWrapper<Article>().in("categoryId", list));
+
+        List<Article> articleList = articleService.list(new QueryWrapper<Article>().in("categoryId", list));
+        if (articleList.size() == 0) {
             return Result.ok();
         }
 
-        return Result.fail();
+        List<Integer> integerList = articleList
+                .stream().map(e -> e.getId()).collect(Collectors.toList());
+
+
+        commentService.remove(new QueryWrapper<Comment>().in("articleid", integerList));
+
+        return Result.ok();
     }
 }
